refactor(admin): break admin subcmds in files
