+++ internal/command/admin/admin.go
@@ -16,95 +16,12 @@
 // under the License.
 package admin
 
-import (
-	"context"
-	"fmt"
+type Admin struct {
+	Dev dev `cmd:"" help:"deploy locally with 1 core and 1 worker docker containers"`
 
-	"github.com/docker/docker/client"
-	"github.com/funlessdev/fl-cli/pkg"
-	"github.com/funlessdev/fl-cli/pkg/admin"
-	"github.com/funlessdev/fl-cli/pkg/log"
-)
+	Init  coreInit `cmd:"" help:"todo init subcommand help"`
+	Join  join     `cmd:"" help:"todo join subcommand help"`
+	Token token    `cmd:"" help:"todo token subcommand help"`
 
-type (
-	Admin struct {
-		Deploy deploy `cmd:"" help:"deploy 1 core and 1 worker locally with docker containers"`
-		Reset  reset  `cmd:"" help:"removes the deployment of local containers"`
-	}
-
-	deploy struct{}
-	reset  struct{}
-)
-
-func (d *deploy) Run(ctx context.Context, logger log.FLogger) error {
-	logger.Info("Deploying funless locally...\n")
-
-	logger.StartSpinner("Setting things up...")
-	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion("1.41"))
-	if err != nil {
-		return err
-	}
-	deployer := admin.NewLocalDeployer(ctx, cli, "fl_net")
-
-	if err := logger.StopSpinner(deployer.Apply(admin.SetupFLNetwork)); err != nil {
-		return err
-	}
-
-	logger.StartSpinner(fmt.Sprintf("pulling Core image (%s) 📦", pkg.FLCore))
-	if err := logger.StopSpinner(deployer.Apply(admin.PullCoreImage)); err != nil {
-		return err
-	}
-
-	logger.StartSpinner(fmt.Sprintf("pulling Worker image (%s) 🗃", pkg.FLWorker))
-	if err := logger.StopSpinner(deployer.Apply(admin.PullWorkerImage)); err != nil {
-		return err
-	}
-
-	logger.StartSpinner("starting Core container 🎛️")
-	if err := logger.StopSpinner(deployer.Apply(admin.StartCoreContainer)); err != nil {
-		return err
-	}
-
-	logger.StartSpinner("starting Worker container 👷")
-	if err := logger.StopSpinner(deployer.Apply(admin.StartWorkerContainer)); err != nil {
-		return err
-	}
-
-	logger.Info("\nDeployment complete!")
-	logger.Info("You can now start using Funless! 🎉")
-
-	return err
-}
-
-func (r *reset) Run(ctx context.Context, logger log.FLogger) error {
-	logger.Info("Removing local funless deployment...\n")
-
-	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion("1.41"))
-	if err != nil {
-		return err
-	}
-
-	logger.StartSpinner("Removing Core container... ☠️")
-	if err := logger.StopSpinner(admin.RemoveFLContainer(ctx, cli, "fl-core")); err != nil {
-		return err
-	}
-
-	logger.StartSpinner("Removing Worker container... 🔪")
-	if err := logger.StopSpinner(admin.RemoveFLContainer(ctx, cli, "fl-worker")); err != nil {
-		return err
-	}
-
-	logger.StartSpinner("Removing the function containers... 🔫")
-	if err := logger.StopSpinner(admin.RemoveFunctionContainers(ctx, cli)); err != nil {
-		return err
-	}
-
-	logger.StartSpinner("Removing fl_net network... ✂️")
-	if err := logger.StopSpinner(admin.RemoveFLNetwork(ctx, cli, "fl_net")); err != nil {
-		return err
-	}
-
-	logger.Info("\nAll clear! 👍")
-
-	return err
+	Reset reset `cmd:"" help:"removes the deployment of local containers"`
 }
